[MODORDERS-1434] Batch create pieces: save po line once
